# read data written with partial flush