see discussion #993 

benchstat 

```
goos: darwin
goarch: arm64
pkg: github.com/klauspost/compress/flate
                          │   old.txt   │              new.txt               │
                          │   sec/op    │   sec/op     vs base               │
DecodeDigitsSpeed1e4-8      44.99µ ± 1%   45.30µ ± 0%  +0.67% (p=0.029 n=10)
DecodeDigitsSpeed1e5-8      444.5µ ± 3%   444.9µ ± 1%       ~ (p=0.971 n=10)
DecodeDigitsSpeed1e6-8      4.537m ± 3%   4.522m ± 2%       ~ (p=0.481 n=10)
DecodeDigitsDefault1e4-8    42.85µ ± 1%   43.32µ ± 1%  +1.11% (p=0.003 n=10)
DecodeDigitsDefault1e5-8    422.5µ ± 1%   424.3µ ± 1%       ~ (p=0.190 n=10)
DecodeDigitsDefault1e6-8    4.278m ± 1%   4.271m ± 1%       ~ (p=0.796 n=10)
DecodeDigitsCompress1e4-8   42.58µ ± 0%   42.70µ ± 1%       ~ (p=0.315 n=10)
DecodeDigitsCompress1e5-8   449.1µ ± 0%   451.3µ ± 1%  +0.49% (p=0.043 n=10)
DecodeDigitsCompress1e6-8   4.523m ± 1%   4.511m ± 2%       ~ (p=0.971 n=10)
DecodeTwainSpeed1e4-8       37.36µ ± 2%   38.27µ ± 1%  +2.41% (p=0.001 n=10)
DecodeTwainSpeed1e5-8       434.7µ ± 0%   436.3µ ± 0%  +0.36% (p=0.035 n=10)
DecodeTwainSpeed1e6-8       4.525m ± 1%   4.491m ± 1%  -0.73% (p=0.011 n=10)
DecodeTwainDefault1e4-8     36.23µ ± 4%   37.83µ ± 5%  +4.42% (p=0.035 n=10)
DecodeTwainDefault1e5-8     403.8µ ± 1%   403.8µ ± 1%       ~ (p=0.280 n=10)
DecodeTwainDefault1e6-8     4.188m ± 1%   4.167m ± 1%       ~ (p=0.393 n=10)
DecodeTwainCompress1e4-8    35.11µ ± 2%   35.51µ ± 2%       ~ (p=0.123 n=10)
DecodeTwainCompress1e5-8    400.1µ ± 1%   398.6µ ± 1%       ~ (p=0.247 n=10)
DecodeTwainCompress1e6-8    4.144m ± 1%   4.124m ± 1%       ~ (p=0.105 n=10)
DecodeRandomSpeed1e4-8      210.6n ± 4%   214.9n ± 2%       ~ (p=0.101 n=10)
DecodeRandomSpeed1e5-8      1.364µ ± 1%   1.381µ ± 1%  +1.21% (p=0.001 n=10)
DecodeRandomSpeed1e6-8      22.11µ ± 1%   21.88µ ± 1%  -1.06% (p=0.000 n=10)
geomean                     193.2µ        194.2µ       +0.50%

                          │   old.txt    │               new.txt               │
                          │     B/s      │     B/s       vs base               │
DecodeDigitsSpeed1e4-8      212.0Mi ± 1%   210.5Mi ± 0%  -0.67% (p=0.029 n=10)
DecodeDigitsSpeed1e5-8      214.5Mi ± 3%   214.4Mi ± 1%       ~ (p=0.971 n=10)
DecodeDigitsSpeed1e6-8      210.2Mi ± 3%   210.9Mi ± 2%       ~ (p=0.481 n=10)
DecodeDigitsDefault1e4-8    222.6Mi ± 1%   220.1Mi ± 1%  -1.09% (p=0.003 n=10)
DecodeDigitsDefault1e5-8    225.7Mi ± 1%   224.8Mi ± 1%       ~ (p=0.184 n=10)
DecodeDigitsDefault1e6-8    222.9Mi ± 1%   223.3Mi ± 1%       ~ (p=0.754 n=10)
DecodeDigitsCompress1e4-8   224.0Mi ± 0%   223.3Mi ± 1%       ~ (p=0.315 n=10)
DecodeDigitsCompress1e5-8   212.3Mi ± 0%   211.3Mi ± 1%  -0.49% (p=0.043 n=10)
DecodeDigitsCompress1e6-8   210.8Mi ± 1%   211.4Mi ± 1%       ~ (p=0.971 n=10)
DecodeTwainSpeed1e4-8       255.2Mi ± 2%   249.2Mi ± 1%  -2.36% (p=0.001 n=10)
DecodeTwainSpeed1e5-8       219.4Mi ± 0%   218.6Mi ± 0%  -0.37% (p=0.030 n=10)
DecodeTwainSpeed1e6-8       210.8Mi ± 1%   212.3Mi ± 1%  +0.74% (p=0.011 n=10)
DecodeTwainDefault1e4-8     263.2Mi ± 4%   252.1Mi ± 5%  -4.23% (p=0.034 n=10)
DecodeTwainDefault1e5-8     236.2Mi ± 1%   236.2Mi ± 1%       ~ (p=0.280 n=10)
DecodeTwainDefault1e6-8     227.7Mi ± 1%   228.9Mi ± 1%       ~ (p=0.382 n=10)
DecodeTwainCompress1e4-8    271.6Mi ± 2%   268.6Mi ± 2%       ~ (p=0.123 n=10)
DecodeTwainCompress1e5-8    238.3Mi ± 1%   239.2Mi ± 1%       ~ (p=0.247 n=10)
DecodeTwainCompress1e6-8    230.1Mi ± 1%   231.2Mi ± 1%       ~ (p=0.109 n=10)
DecodeRandomSpeed1e4-8      44.22Gi ± 4%   43.33Gi ± 2%       ~ (p=0.105 n=10)
DecodeRandomSpeed1e5-8      68.28Gi ± 1%   67.48Gi ± 1%  -1.17% (p=0.001 n=10)
DecodeRandomSpeed1e6-8      42.12Gi ± 1%   42.57Gi ± 1%  +1.07% (p=0.000 n=10)
geomean                     493.7Mi        491.2Mi       -0.50%
```