@vajexal 2024-08-31
> Please go fmt:
> 
> ```
> 1,14d0
> < diff flate/inflate.go.orig flate/inflate.go
> < --- flate/inflate.go.orig
> < +++ flate/inflate.go
> < @@ -302,[8](https://github.com/klauspost/compress/actions/runs/10645790779/job/29511975229?pr=996#step:4:9) +302,8 @@
> <  type flushMode uint8
> <  
> <  const (
> < -	syncFlush flushMode = iota // return data after sync flush block
> < -	partialFlush // return data after each block
> < +	syncFlush    flushMode = iota // return data after sync flush block
> < +	partialFlush                  // return data after each block
> <  )
> <  
> <  // Decompress state.
> ```

yeah, i forgot. Done